Data Collection Method writted
Written a method to collect all of the user associated data within the login function class. Will be called once login in successful, returns a list of all of the populated rows in User_Societies table relevant to that user.

Will need updating with a hashmap to get the website address in the same result